Internals

String references

The Strref type mirrors the type used in-game:

InfinityEngine.Strref — Type

Strref

Index (32-bit) referring to a translated string in a "dialog.tlk" file.

These values are zero-indexed (Julia arrays are generally 1-indexed). Index 0 is the string "<NO TEXT>", which is technically valid, but should never be actually used in-game; we use this as a marker for invalid or missing strings, which allows functions to gracefully fail by never referring to an inexistent string.

Resource references

The game admits several resource types (about 40 of them in total), for referring to e.g. items, scripts or spells.

This module uses two systems for referring to game resources:

the game's own key type (“shortrefs”): 8-byte strings, used as either a key in the chitin.key file or as a file name in the filesystem;

user-friendly references (Resref), which can be arbitrary strings and are namespaced to prevent name collision between mods.

This double system should be transparent to mod authors; this is a strong design goal of this module. To this end, the Game structure maintains a dictionary for Resref to shortref conversion. This dictionary is stored in the shortref field of the GameResources structure, and is saved in the [resources] section of the "state.toml" file.

InfinityEngine.Resref — Type

Resref{T}

A (long) resource descriptor: this contains a (static) type and a (dynamic) namespace and name uniquely identifying the resource.

InfinityEngine.ResIO — Type

ResIO{Type}, ResIO"TYPE"

An IO object marked (statically) with the type of resource being read (either from filesystem or from a BIF content), as well as (dynamically) with global properties of the resource (i.e. resource name, root object).

Since the resource type determines the read/write methods (and only takes a finite set of values) this is a type parameter.

The resource name (i.e. basename of the file without extension) is stored as a field. The name is canonicalized as upper-case.

Defined methods include:

ResIO"EXT": macro defining static value for this file type.

While resource names (either 8-byte or long) are dynamic properties, resource type (“item”/“script” etc.) is a static, compile-time property; thus it is implemented as a type parameter instead of as a structure field. This allows handling resources via generic functions, which is the best of both worlds:

it is as fast as writing a big list of similar functions find_item, find_creature etc.;

it is as readable/maintenable as writing a single function find_resource(type, name).

Thanks Julia for generic functions!

The Game structure

This structure holds global information about a particular game installation. It contains sub-structures dedicated to specific information: GameResources, GameStrings, DialogContext.

InfinityEngine.Game — Type

Game

Main structure holding all top-level data for a game installation, including:

resource repository (three kinds: key/bif, override, memory database of modified resources);

tlk strings,

conversion of resource references to 8-byte short references;

dialog-building context.

This structure works as a pseudo-dictionary: indexing it with keys of a given Resref type returns data structures of the corresponding resource type.

game[resref]: returns the data structure described by this resource.

get and get! methods, e.g. get(game, resref) do ... end

keys(game, type): returns a vector of all names of existing resources of this type.

GameResources

This structure is responsible for accessing game resources from their symbolic references, as well as for saving them.

InfinityEngine.GameResources — Type

GameResources

A structure holding all the game resources, stored in either key/bif files, the override directory, or in memory.

This structure is also responsible for converting between these storage types, i.e. reading the resources from the filesystem or committing the in-memory storage to the filesystem.

It stores mirrors of the chitin.key and override filesystem tables, as well as the conversion from resource references to short in-game references, and the table of modified resources.

The table of modified resources is the set of all game resources needing to be written to the filesystem in the next commit call.

RootedResource

This is the common subtype for all user-modifiable resources (except dialogs, which have their own syntax). The resources all know their own reference (via the ref field) and child structures (e.g. item abilities) store a pointer to the root resource (in this case, a pointer back to the item).

The setproperty! function is overloaded so that, whenever a field of such a structure is modified, the modified object will be stored in the appropriate table in the GameResource object. Then the commit function knows which structures need to be saved to the override/ directory.

InfinityEngine.RootResource — Type

RootResource

The root of a resource tree, i.e. a resource which will be saved in a game file. This resource also holds an identifier designating it (and the game file).

InfinityEngine.RootedResource — Type

RootedResource

A game resource holding a reference to the root resource (i.e. the one which will eventually be saved in a game file).

The following methods needs to be defined:

root(x): returns the root object

root!(x, r): sets the root object

register!(root): marks this root object as modified.

TODO: it would theoretically be possible to implement ownership for resources (or even individual fields); this would completely solve mod conflicts (but be quite hard to implement).

GameStrings

This structure manages translation of game strings. It stores a dictionary between newly added game strings (in untranslated, source form) and in-game numeric Strrefs.

InfinityEngine.GameStrings — Type

GameStrings

Collection of game strings, indexed by string and language.

Interface

Strref(gamestrings, x): string key to Strref conversion.

gamestrings[language, strref]: Strref to String conversion.

commit(gamestrings): saves state to filesystem.

init!(directory): loads state from filesystem.

The state file

All persistent information from the Game structure is stored in this file.

This file has sections [resources] and [strings] corresponding to the sub-structures of Game.